requirements.txt
absl-py==2.1.0
anndata==0.10.8
array_api_compat==1.8
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
grpcio==1.65.5
h5py==3.11.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
legacy-api-wrap==1.4
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
protobuf==5.27.3
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
scanpy==1.10.2
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
session_info==1.0.0
setuptools==72.1.0
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
tzdata==2024.1
umap-learn==0.5.6
Werkzeug==3.0.3
wheel==0.43.0